created exhibition seed file, exhibition index show all route working w/ seed data
