+++ package.json
@@ -1,20 +1,21 @@
 {
-  "name": "express-auth-mongoose-boilerplate",
+  "name": "art-exhibition-api",
   "version": "1.0.0",
-  "description": "Auth boilerplate for mongoose express using token auth",
+  "description": "art collecting api",
   "main": "server.js",
   "scripts": {
     "start": "nodemon server.js",
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "seed": "node ./app/models/seed.js"
   },
-  "author": "Timm Schoenborn",
+  "author": "The Con Artist",
   "license": "ISC",
   "dependencies": {
     "axios": "^1.3.3",
     "bcrypt": "^5.0.1",
     "cors": "^2.8.5",
     "dotenv": "^10.0.0",
-    "express": "^4.17.1",
+    "express": "^4.18.2",
     "jsonwebtoken": "^8.5.1",
     "method-override": "^3.0.0",
     "mongodb": "^4.2.1",
