+++ server.js
@@ -7,6 +7,7 @@ const cors = require('cors')
 const exampleRoutes = require('./app/routes/example_routes')
 const userRoutes = require('./app/routes/user_routes')
 const artworkRoutes = require('./app/routes/artwork_routes')
+const exhibitionRoutes = require('./app/routes/exhibition_routes')
 
 // require middleware
 const errorHandler = require('./lib/error_handler')
@@ -69,6 +70,8 @@ app.use(requestLogger)
 app.use(exampleRoutes)
 app.use(userRoutes)
 app.use(artworkRoutes)
+app.use(exhibitionRoutes)
+
 
 // register error handling middleware
 // note that this comes after the route middlewares, because it needs to be
